client/requirements.txt
transitions
paho-mqtt